Move components to design system
Migrate MullvadButton, Feature indicator and filter chip to design
system
