+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/compose/button/TestMethodButton.kt
@@ -7,6 +7,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import net.mullvad.mullvadvpn.R
 import net.mullvad.mullvadvpn.compose.preview.TestMethodButtonPreviewParameterProvider
+import net.mullvad.mullvadvpn.lib.ui.designsystem.PrimaryButton
 import net.mullvad.mullvadvpn.lib.ui.theme.AppTheme
 
 @Preview
